fix: 메인페이지 분리
