+++ pages/main/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./github-server";
+export * from "./loader";
