+++ pages/main/index.ts
@@ -1,0 +1,2 @@
+export { loader } from "./api";
+export { MainPage } from "./ui/main-page";
